Refactor action list management.
The action lists were duplicated all over the place.  This commit
removes some of the duplication by introducing a new module-level
function named `get_application_actions` to the model modules.  This
function returns a list of action dicts.  Since it returns a list, the
`*actions` parameter to `serialize_model_instance` has been converted
to a keyword parameter so that we can avoid splatting lists everywhere.